More changes to master manual paw data script
- removed date field in RatData structure
- Added cell array for determining if dominant paw used or not
- Added AutomaticTriggerFrame and AutomaticPeakFrame variables to
RatData.Videos, which will be filled by results of identifyTriggerFrame
- If ROI produces peak frame > 5 away from manually determined start
frame, it will prompt the user to change the ROI by re-drawing a
rectangle in the appropriate mirror. ROI changes are carried over to
subsequent trials, based on the assumption that over time it will be
closest to the optimal ROI. Resets for every day of trials.
- Paw preference changes according to what is documented in rat's .csv
file on NAS
- Added fields to RatData.VideoFiles to output ROI used and paw
preference. Accuracy also calculated for date overall (automatically
determined peak frame considered accurate if <5 away from manually
determined start frame) and displayed in RatData structure.