My scrap script for testing identifyTriggerFrame
- exports manual start frame, automatic trigger frame, automatic peak
frame, and ROI used data to excel file titled
'R0027_Start_Frame_Data.xlsx'. If the manual start frame or
automatically determined peak frame is NaN, it makes it a 0 in Excel
- tests first differential thresholds from 200 to 300 with an interval
of 10 using data from RatData(27).VideoFiles.mat. Calculates accuracy
for each video/threshold combination (480 in all).
- Calculates overall accuracy: gives score of 1 if automatically
determined trigger frame or peak frame is within 5 frames of manually
determined start frame, gives score of 0 if both are >5 frames away, and
gives score of .5 if one is accurate but other is not.
- Also calculates accuracy for trigger frame and peak frame separately:
1 = both <5 frames or just one, 0 = both >5 frames or opposite frame is
accurate but the one in question is not.
- calculates and plots mean accuracy, accuracy for trigger frame and
peak frame separately, and differentials of all the above in a single
plot
- exports all the raw data (automatically determined trigger frames and
peak frames for each video/threshold combo + data from
RatData(27).VideoFiles.mat.) and accuracy data plotted above to Excel
file 'R0027_Start_Frame_Data_With_Highlights.xlsx'