Interval changed
Changed frames analyzed from the first 10 + every 5 after that, up to 40
frames after the start frame. Now just does the start frame and every 8
frames after that, for 5 frames for analysis total. This way we can get
a rough measure of paw movement that can be presented at lab meeting.